Lazy-load workspace route editor code
