+++ apps/app/bundle-budget.json
@@ -14,10 +14,9 @@
     "them when a change wins headroom. Raising one is a deliberate decision",
     "that needs a reason in the pull request, not a routine edit.",
     "",
-    "There is deliberately no per-chunk limit. Parse and compile cost scales",
-    "with the total bytes on the boot path, not with how they are divided, so",
-    "splitting one boot chunk into two would satisfy a per-chunk cap while",
-    "making the page slightly slower for the extra request.",
+    "The workspace checkout display chunk has a separate limit because it was",
+    "the shared owner of the prompt, Markdown, and diff renderers. The route",
+    "package check keeps those features behind their dynamic boundaries.",
     "",
     "forbiddenBootPackages is the more important half. These packages are all",
     "large and all needed only after a user action (open a diff, focus the",
@@ -27,6 +26,7 @@
   ],
   "maxBootBytes": 1707047,
   "maxBootBrotliBytes": 445742,
+  "maxWorkspaceCheckoutDisplayChunkBytes": 512000,
   "forbiddenBootPackages": [
     "@pierre/diffs",
     "@pierre/theming",
@@ -41,6 +41,27 @@
     "@tiptap/react",
     "@tiptap/starter-kit",
     "@xterm/xterm",
+    "katex",
+    "oniguruma-to-es",
+    "prosemirror-model",
+    "prosemirror-state",
+    "prosemirror-view",
+    "rehype-katex",
+    "shiki"
+  ],
+  "forbiddenWorkspaceRoutePackages": [
+    "@pierre/diffs",
+    "@pierre/theming",
+    "@shikijs/core",
+    "@shikijs/engine-javascript",
+    "@shikijs/engine-oniguruma",
+    "@shikijs/langs",
+    "@shikijs/vscode-textmate",
+    "@tiptap/core",
+    "@tiptap/pm",
+    "@tiptap/react",
+    "@tiptap/starter-kit",
+    "cytoscape",
     "katex",
     "oniguruma-to-es",
     "prosemirror-model",
