+++ apps/app/src/components/secondary-panel/FilePreview.test.tsx
@@ -13,7 +13,7 @@ import {
   FilePreview,
   buildCsvPreviewData,
   getCsvTruncationNote,
-} from "./FilePreview";
+} from "./FilePreviewImpl";
 import { SecondaryPanelFilePreview } from "./ThreadStorageFilePreview";
 
 interface MockPierreFileProps {
@@ -401,7 +401,7 @@ describe("FilePreview", () => {
     expect(getCsvTruncationNote(preview, preview.rows.length - 1)).toBeNull();
   });
 
-  it("uses the CSV table preview for loaded CSV text files", () => {
+  it("uses the CSV table preview for loaded CSV text files", async () => {
     render(
       <SecondaryPanelFilePreview
         activePath="exports/scores.csv"
@@ -418,7 +418,7 @@ describe("FilePreview", () => {
     );
 
     expect(
-      screen.getByRole("table", { name: "scores.csv CSV preview" }),
+      await screen.findByRole("table", { name: "scores.csv CSV preview" }),
     ).not.toBeNull();
     expect(screen.getByRole("cell", { name: "Ada" })).not.toBeNull();
     expect(screen.getByRole("cell", { name: "10" })).not.toBeNull();
