+++ apps/app/src/components/secondary-panel/threadSecondaryPanelLayout.ts
@@ -1,0 +1,4 @@
+// The split-workspace empty-state panel uses the same resize bounds as the
+// full secondary panel.
+export const THREAD_SECONDARY_PANEL_MIN_SIZE_PERCENT = 24;
+export const THREAD_SECONDARY_PANEL_MAX_SIZE_PERCENT = 70;
