+++ apps/app/src/views/PluginPanelView.tsx
@@ -1,24 +1,14 @@
 import { useParams } from "react-router-dom";
-import { WorkerPoolContextProvider } from "@pierre/diffs/react";
 import { PageShell } from "@/components/ui/page-shell.js";
 import { EmptyStatePanel } from "@bb/shared-ui/empty-state";
 import { PluginSlotMount } from "@/components/plugin/PluginSlotMount";
-import {
-  createDiffWorker,
-  getDiffWorkerPoolSize,
-} from "@/lib/diff-worker-pool";
 import { usePluginSlots } from "@/lib/plugin-slots";
+import { ThreadDetailWorkerPoolProvider } from "./thread-detail/ThreadDetailWorkerPoolProvider";
 
 // Plugins can render `@pierre/diffs` FileDiff (the specifier is shimmed to
 // the host's copy); syntax highlighting needs a worker pool in React context.
 // Thread panes get theirs from the split workspace — standalone nav panels get
 // one here.
-const WORKER_POOL_OPTIONS = {
-  workerFactory: createDiffWorker,
-  poolSize: getDiffWorkerPoolSize(),
-};
-const HIGHLIGHTER_OPTIONS = {};
-
 /**
  * The route surface for plugin `navPanel` slots (plugin design §5.2):
  * /plugins/:pluginId/:panelPath renders the matching registered panel
@@ -80,17 +70,9 @@ export function PluginPanelView(props: PluginPanelViewProps = {}) {
   );
   // The provider spawns workers eagerly; environments without Worker
   // (jsdom tests) just render diffs unhighlighted.
-  const mount =
-    typeof Worker === "undefined" ? (
-      slotMount
-    ) : (
-      <WorkerPoolContextProvider
-        poolOptions={WORKER_POOL_OPTIONS}
-        highlighterOptions={HIGHLIGHTER_OPTIONS}
-      >
-        {slotMount}
-      </WorkerPoolContextProvider>
-    );
+  const mount = (
+    <ThreadDetailWorkerPoolProvider>{slotMount}</ThreadDetailWorkerPoolProvider>
+  );
 
   // Full-bleed: the negative margins undo the app layout's `p-4 md:p-5`
   // route padding. Plugins opt into their own padding and scrolling.
