+++ apps/app/src/views/thread-detail/ThreadDetailWorkerPoolProviderImpl.tsx
@@ -1,0 +1,30 @@
+import { WorkerPoolContextProvider } from "@pierre/diffs/react";
+import type { ReactNode } from "react";
+import {
+  createDiffWorker,
+  getDiffWorkerPoolSize,
+} from "@/lib/diff-worker-pool";
+
+const WORKER_POOL_OPTIONS = {
+  workerFactory: createDiffWorker,
+  poolSize: getDiffWorkerPoolSize(),
+};
+const HIGHLIGHTER_OPTIONS = {};
+
+export function ThreadDetailWorkerPoolProvider({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  if (typeof Worker === "undefined") {
+    return children;
+  }
+  return (
+    <WorkerPoolContextProvider
+      poolOptions={WORKER_POOL_OPTIONS}
+      highlighterOptions={HIGHLIGHTER_OPTIONS}
+    >
+      {children}
+    </WorkerPoolContextProvider>
+  );
+}
